Moved error controller and moved error model
