+++ src/main/java/org/seattlevoluntech/controllers/ErrorController.java
@@ -3,8 +3,10 @@ package org.seattlevoluntech.controllers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.web.ErrorProperties;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
@@ -15,6 +17,11 @@ public class ErrorController{
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public static class ErrorNotFound extends RuntimeException {
+        public ErrorNotFound(String exception) {
+            super(exception);
+        }
+    }
 
 }
